Fix pulling in items from other revisions
The presswise data is odd - it relates the quote item by the quote id + revision instead of the unique id on the quote.
Adds in a condition to only pull in items for the imported revision (it's always 1 in our case but using the revision from the
quote is more correct).